Init logging to file
