+++ node-daemon/src/main.rs
@@ -15,8 +15,6 @@
 
 pub async fn run() -> anyhow::Result<()> {
     let opts = node_lib::Options::from_args(std::env::args_os());
-    logging::init_logging();
-    logging::log::info!("Command line options: {opts:?}");
     let setup_result = node_lib::setup(opts).await?;
     match setup_result {
         node_lib::NodeSetupResult::Node(node) => {
