+++ utils/Cargo.toml
@@ -32,6 +32,7 @@ criterion.workspace = true
 regex.workspace = true
 rstest.workspace = true
 static_assertions.workspace = true
+tempfile.workspace = true
 tracing.workspace = true
 tokio = { workspace = true, features = ["macros"] }
